+++ .changeset/sixty-pandas-hide.md
@@ -1,0 +1,5 @@
+---
+'@embr-modules/periscope': patch
+---
+
+Resolve NPE related to `ClientResourceManager` initialization for `ClientResourceWorkspace`'s Welcome Page.
