+++ modules/periscope/designer/src/main/kotlin/com/mussonindustrial/ignition/embr/periscope/resources/ClientResourceWorkspace.kt
@@ -7,6 +7,10 @@ import com.inductiveautomation.ignition.designer.model.DesignerContext
 import com.inductiveautomation.ignition.designer.navtree.model.AbstractNavTreeNode
 import com.inductiveautomation.ignition.designer.navtree.model.MutableNavTreeNode
 import com.inductiveautomation.ignition.designer.tabbedworkspace.*
+import com.inductiveautomation.ignition.designer.workspacewelcome.RecentlyModifiedTablePanel
+import com.inductiveautomation.ignition.designer.workspacewelcome.ResourceBuilderDelegate
+import com.inductiveautomation.ignition.designer.workspacewelcome.ResourceBuilderPanel
+import com.inductiveautomation.ignition.designer.workspacewelcome.WorkspaceWelcomePanel
 import com.mussonindustrial.ignition.embr.periscope.icons.PeriscopeIcons
 import com.mussonindustrial.ignition.embr.periscope.navtree.model.ClientResourceRootFolder
 import java.util.*
@@ -58,9 +62,33 @@ class ClientResourceWorkspace(
 
     override fun createWorkspaceHomeTab(): Optional<JComponent> {
         return Optional.of<JComponent>(
-            ClientResourceWorkspaceHomeTab(context, clientResourceManager, this)
+            object :
+                WorkspaceWelcomePanel(BundleUtil.i18n("periscope.client-resource.nouns-long")) {
+                override fun createPanels(): List<JComponent> {
+                    return listOf<JComponent>(
+                        ResourceBuilderPanel(
+                            context,
+                            BundleUtil.i18n("periscope.client-resource.noun"),
+                            ClientResource.type.rootPath(),
+                            clientResourceManager.descriptors.map {
+                                it.asResourceBuilderDelegate()
+                            },
+                            this@ClientResourceWorkspace::open,
+                        ),
+                        RecentlyModifiedTablePanel(
+                            context,
+                            ClientResource.type,
+                            BundleUtil.i18n("periscope.client-resource.nouns-long"),
+                            this@ClientResourceWorkspace::open,
+                        ),
+                    )
+                }
+            }
         )
     }
+
+    fun ClientResourceDescriptor<*>.asResourceBuilderDelegate(): ResourceBuilderDelegate =
+        ResourceBuilderDelegate.build(nounLongKey, icon, createEmpty()::applyToBuilder)
 
     class NewClientResourceAction(
         workspace: TabbedResourceWorkspace,
